+++ resolwe/flow/executors/docker/prepare.py
@@ -54,3 +54,9 @@ class FlowExecutorPreparer(BaseFlowExecutorPreparer):
             return constants.UPLOAD_VOLUME
 
         return os.path.join(constants.UPLOAD_VOLUME, filename)
+
+    def get_environment_variables(self):
+        """Return dict of environment variables that will be added to executor."""
+        return {
+            'TMPDIR': os.path.join(constants.DATA_VOLUME, '.tmp')
+        }