+++ resolwe/flow/tests/test_env_vars.py
@@ -5,11 +5,12 @@ from django.conf import settings
 
 from resolwe.flow.managers import manager
 from resolwe.flow.models import Data, Process
-from resolwe.test import TransactionTestCase
+from resolwe.test import TransactionTestCase, with_docker_executor
 
 
 class EnvVarsTest(TransactionTestCase):
 
+    @with_docker_executor
     def test_envvars(self):
         flow_executor = copy.copy(getattr(settings, 'FLOW_EXECUTOR', {}))
         flow_executor['SET_ENV'] = {
@@ -26,12 +27,14 @@ class EnvVarsTest(TransactionTestCase):
                 output_schema=[
                     {'name': 'resolweapihost', 'type': 'basic:string:'},
                     {'name': 'setenvtest', 'type': 'basic:string:'},
+                    {'name': 'tmpdir', 'type': 'basic:string:'},
                 ],
                 run={
                     'language': 'bash',
                     'program': """
 re-save resolweapihost $RESOLWE_HOST_URL
 re-save setenvtest $SET_ENV_TEST
+re-save tmpdir $TMPDIR
 """
                 }
             )
@@ -48,3 +51,4 @@ re-save setenvtest $SET_ENV_TEST
 
             self.assertEqual(data.output['resolweapihost'], 'some.special.host')
             self.assertEqual(data.output['setenvtest'], 'test_var')
+            self.assertEqual(data.output['tmpdir'], '/data/.tmp')