+++ docs/CHANGELOG.rst
@@ -19,6 +19,8 @@ Added
 -----
 - Raise an exception in Docker executor if container doesn't start for 30
   seconds
+- Set ``TMPDIR`` environment variable in Docker executor to ``.tmp`` dir in
+  data directory to prevent filling up container's local storage
 
 Fixed
 -----